Upload of submitted code for class project.
Update:
driver.cpp with an extremely naive algorithm due to extremely costly communication times between processes.

Upload:
instructions.pdf - The instructions for the project.
report.docx and run_times.xlsx - The report written per the instructions and the run times attained for different runs of the program.
readme.txt - The readme for the grader.
proj3-script - The execution script used for grading.